// src/stores/dedicated-resources/models/IDedicatedResourcesState.ts
export default interface IDedicatedResourcesState {
  readonly listData: DedicatedResourcesEnvelope;
  readonly listDataBulkUpdate: DedicatedResourcesEnvelope;
  readonly listHistoryContract: GetHistoryContractEnvelope;
  readonly listActivityReport: GetActivityReportEnvelope;
  readonly listDeductions: DeductionsEnvelope;
  readonly listOtherBenefit: OtherBenefitEnvelope;
  readonly listSalaryBenefit: SalaryBenefitEnvelope;
  readonly listOtherLastContract: OtherBenefitLastContractEnvelope;
  readonly listOtherTemplateProject: OtherBenefitTemplateProject[];
  readonly DeductionById: DeductionsModel;
  readonly SalaryBenefitById: SalaryBenefitModelById;
  readonly DropdownContractStatus: DropdownDedicatedResourcesModel[];
  readonly DropdownDepartment: DropdownDedicatedResourcesModel[];
  readonly DropdownSupervisor: DropdownDedicatedResourcesModel[];
  readonly DropdownDeductionType: DropdownDedicatedResourcesModel[];
  readonly DropdownDeductionDesc: DropdownDedicatedResourcesModel[];
  readonly DropDownOtherBenefitType: DropdownDedicatedResourcesModel[];
  readonly DropDownOtherBenefitDesc: DropdownDedicatedResourcesModel[];
  readonly DropDownSalaryBenefitType: DropdownDedicatedResourcesModel[];
  readonly DropDownSalaryBenefitDesc: DropdownDedicatedResourcesModel[];
  readonly DropDownUMRYearSalaryBenefit: DropdownDedicatedResourcesModel[];
  readonly DropDownUMRSalaryBenefit: DropdownDedicatedResourcesModel[];
  readonly DropdownSenderAdmin: DropdownDedicatedResourcesModel[];
  readonly DropDownBuCost: DropdownDedicatedResourcesModel[];
  readonly DropdownEmployee: EmployeeModel[];
  readonly TakeHomePay: DropdownDedicatedResourcesModel;
  readonly CurrentSalary: DropdownDedicatedResourcesModel;
  readonly WorkFlowHeader: WorkFlowHeaderEnvelope;
  readonly ListApprovalButton: ListApprovalModel[];
  readonly VerificationDataStatus: VerificationDataStatusModel[];
  readonly SearchSOOI: GetSearchSOOIModel[];
  readonly EmployeeInfo: GetEmployeeInfoModel;
  readonly EmployeeDetail: GetEmployeeDetailModel;
  readonly ProjectInfo: GetProjectInfoModel;
  readonly DocumentTracking: DocumentTrackingModel;
  readonly CheckLastEmployeeContract: ResultCheck;
  readonly CheckSubmited: ResultCheck;
  readonly CheckApproval: ResultCheck;
  readonly error: boolean;
  readonly refreshPage: boolean;
  readonly resultActions: ResultActions;
  readonly isExportExcel: boolean;
}